requirements.txt
kivy==2.3.1
kivymd==1.2.0
requests==2.32.3
python-dateutil==2.9.0.post0
androidstorage4kivy==0.1.3
